post va home

+++ post_service/src/main/java/com/xdpmtmhpl/post_service/repository/PostRepository.java
@@ -1,13 +1,14 @@
 package com.xdpmtmhpl.post_service.repository;
 
-import com.xdpmtmhpl.post_service.models.Post;
+import com.xdpmtmhpl.post_service.model.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
 @Repository
-public interface PostRepository extends JpaRepository<Post, Long> {
-    List<Post> findByUserId(Long userId);
+public interface PostRepository extends JpaRepository<Post, Integer> {
+    List<Post> findByUserId(Integer userId);
 }
+
